requirements.txt
customtkinter>=5.2.0
requests>=2.31.0
aiohttp>=3.9.0
yt-dlp
Pillow>=10.0.0
python-dotenv>=1.0.0
packaging>=23.0
